moved all interfaces to types.mts, version bumped to 2.0.0
