+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.0.0
+
+- Convert code to TypeScript
+- Convert code building from Rollup to Vite v4
+- Convert unit testing from Jest to Vitest
+
 ## 1.0.0
 
 - Initial package publish
